Generate Adminhtml Quiz grid
- Refactor db_schema
- Using the Magento 2 PHPStorm plugin, generate the Adminhtml Quiz grid
  using the `Magento 2 Entity` selection
- Add a product grid to the UI Component by adapting the following
  tutorial: https://www.rohanhapani.com/magento-2-how-to-add-product-grid-in-ui-form-using-uicomponent/

Previous attempts to add a product grid to the form using the
`insertListing` UI Component inserted a product grid successfully, but
did not submit any sort of form data in a way one could see.

This product grid will need to be refactored in the future, as many of
the blocks used utilize deprecated classes. But, seeing as there are no
great examples of `insertListing` being used in the Magento 2 core, and
this block method of UI Components is still used in the Category page to
select products, the technique should be safe to use for some time.